backend/requirements.txt
# Core Dependencies
Django==5.0.7
asgiref==3.9.1
sqlparse==0.5.3
pytz==2025.2
tzdata==2025.2
redis==5.0.1
daphne==4.0.0
# Django REST Framework
djangorestframework==3.15.1

# Celery Task Queue
celery==5.3.6
kombu==5.5.4
amqp==5.3.1
billiard==4.2.1
vine==5.1.0

# Celery Beat (Scheduler)
django-celery-beat==2.8.1

# Celery Monitoring
-e git+https://github.com/mher/flower.git@d6898881a36945ebab86d6f00f1d2d16dbac7b06#egg=flower

# Database
psycopg2-binary==2.9.10  # PostgreSQL adapter (Python 3.13 compatible)
psutil==7.0.0  # System metrics and process utilities

# Image Processing
Pillow==10.4.0  # Required for ImageField

# Authentication
djangorestframework-simplejwt==5.3.1  # JWT Authentication
django-ipware==7.0.1  # IP address utilities

# API Documentation
drf-yasg==1.21.7  # Yet another Swagger generator

# CORS Headers
django-cors-headers==4.3.1

# WebSocket Support (Django Channels)
channels==4.0.0
channels-redis==4.1.0
daphne==4.0.0  # ASGI server for Channels

# Filtering
django-filter==23.5

# Model Utilities
django-model-utils==4.3.1

# Messaging Dependencies
python-dateutil==2.9.0.post0
reportlab==4.1.0  # For PDF generation in exports

# Environment Variables
python-dotenv==1.0.1

# CLI Tools
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
prompt_toolkit==3.0.51
wcwidth==0.2.13

# Monitoring
prometheus_client==0.22.1

# Utilities
humanize==4.12.3
packaging==25.0
six==1.17.0
tornado==6.5.1

# Build Tools
setuptools==80.9.0
wheel==0.45.1
